rsvn-12: Display dialog on delete

+++ project-data/angular/rsvn/src/app/frontdesk/rsvn-edit/rsvn-edit.component.ts
@@ -8,6 +8,7 @@ import { IGuest } from '@app/_interface/guest';
 import { IRoom } from '@app/_interface/room';
 import { RsvnService } from '@app/_services/rsvn.service';
 import { RoomService } from '@app/_services/room.service';
+import {DangerDialogComponent, ManagerService} from "@app/shared/dialog";
 
 @Component({
   selector: 'app-rsvn-edit',
@@ -22,8 +23,8 @@ export class RsvnEditComponent implements OnInit, OnChanges {
     private systemService: SystemService,
     private authService: AuthService,
     private rsvnService: RsvnService,
-    private roomService: RoomService
-
+    private roomService: RoomService,
+    private dialogManagerService: ManagerService,
   ) { }
 
 
@@ -184,26 +185,31 @@ export class RsvnEditComponent implements OnInit, OnChanges {
 
   //---------------------------------
   deleteRsvn(rsvn: any) {
-
-    if (!this.rsvnLocked()) {
-
-      this.genericService.deleteItem('rsvn', rsvn).subscribe(
-        data => {
-          this.currRsvn = null
-          this.currRsvnChange.emit(this.currRsvn)
-          this.rsvnEditFormInit()
-        },
-        err => {
-          console.log("ERROR in rsvn Delete", err)
+    this.dialogManagerService.openDialog<DangerDialogComponent>(DangerDialogComponent, {
+      width: '250px'
+    }).afterClosed().subscribe(result => {
+      console.log('The dialog was closed', result);
+      if (result) {
+        if (!this.rsvnLocked()) {
+          this.genericService.deleteItem('rsvn', rsvn).subscribe(
+            data => {
+              this.currRsvn = null
+              this.currRsvnChange.emit(this.currRsvn)
+              this.rsvnEditFormInit()
+            },
+            err => {
+              console.log("ERROR in rsvn Delete", err)
+            }
+          )
         }
-      )
-    }
+      }
+    });
   }
   //---------------------------------
   ngOnChanges(changes: SimpleChanges) {
     this.rsvnEditForm.reset()
 //    if (this.currRsvn && this.currRsvn.id == 0) {
-  
+
       this.rsvnEditFormInit()
 //    }
     this.loadRsvn(this.currRsvn)
